+++ makefile/help.py
@@ -18,30 +18,33 @@ along with this program.  If not, see <https://www.gnu.org/licenses/>.
 """
 
 HELP = """
-\033[1;36mFarix Build System\033[0m
-\033[90mhttps://github.com/farismuhammad17/Farix\033[0m
+\x1b[1;36mFarix Build System\x1b[0m
+\x1b[90mhttps://github.com/farismuhammad17/Farix\x1b[0m
 
-\033[90mTIP: Use 'source make.env' to alias make.py as 'm'\033[0m
-Usage: \033[1m m [target] <architecture>\033[0m
+\x1b[90mTIP: Use 'source make.env' to alias make.py as 'm'\x1b[0m
+Usage: \x1b[1m m [target] <architecture>\x1b[0m
 
-\033[1;36mTargets:\033[0m
-  \033[1;32mall\033[0m          Build the kernel and deploy apps
-  \033[1;32mlibc\033[0m         Compile the internal C standard library
-  \033[1;32mdisk\033[0m         Generate the bootable disk image for emulation
-  \033[1;32mget_deps\033[0m     Fetch and verify build-tool dependencies
-  \033[1;32mwipe_usb\033[0m     Wipe every single file inside the Boot USB
-  \033[1;32musb\033[0m          Copy farix.bin into Boot USB \033[90m(provide in make.conf.json)\033[0m
-  \033[1;32mqemu\033[0m         Build and launch in QEMU \033[90m(Fullscreen mode)\033[0m
-  \033[1;32mqemu_\033[0m        Build and launch in QEMU \033[90m(Windowed mode)\033[0m
-  \033[1;32mclean\033[0m        Remove all build artifacts and object files \033[90m(Optionally provide exclusions after; 'apps' to delete only apps)\033[0m
-  \033[1;32mapps\033[0m         Compile and add all apps into disk
-  \033[1;32mcompile_apps\033[0m Compile all the files inside apps into ELF
-  \033[1;32mdeploy_apps\033[0m  Deploy all ELF files inside apps folder
-  \033[1;32mdefs\033[0m         Interactive program to return definitions and usage of any function
-  \033[1;32mlint\033[0m         Lint the code \033[90m(Useful for developing)\033[0m
-  \033[1;32mhelp\033[0m         Display this menu
+\x1b[1;36mTargets:\x1b[0m
+  \x1b[1;32mall\x1b[0m          Build the kernel and deploy apps
+  \x1b[1;32mlibc\x1b[0m         Compile the internal C standard library
+  \x1b[1;32mdisk\x1b[0m         Generate the bootable disk image for emulation
+  \x1b[1;32mget_deps\x1b[0m     Fetch and verify build-tool dependencies
+  \x1b[1;32mwipe_usb\x1b[0m     Wipe every single file inside the Boot USB
+  \x1b[1;32musb\x1b[0m          Copy farix.bin into Boot USB \x1b[90m(provide in make.conf.json)\x1b[0m
+  \x1b[1;32mqemu\x1b[0m         Build and launch in QEMU \x1b[90m(Fullscreen mode)\x1b[0m
+  \x1b[1;32mqemu_\x1b[0m        Build and launch in QEMU \x1b[90m(Windowed mode)\x1b[0m
+  \x1b[1;32mclean\x1b[0m        Remove all build artifacts and object files \x1b[90m(Optionally provide exclusions after; 'apps' to delete only apps)\x1b[0m
+  \x1b[1;32mapps\x1b[0m         Compile and add all apps into disk
+  \x1b[1;32mcompile_apps\x1b[0m Compile all the files inside apps into ELF
+  \x1b[1;32mdeploy_apps\x1b[0m  Deploy all ELF files inside apps folder
+  \x1b[1;32mdefs\x1b[0m         Interactive program to return definitions and usage of any function
+  \x1b[1;32mlint\x1b[0m         Lint the code \x1b[90m(Useful for developing)\x1b[0m
+  \x1b[1;32mhelp\x1b[0m         Display this menu
 
-\033[1;36mArchitectures:\033[0m
-  \033[1mx86_32\033[0m       32-bit x86 \033[90m(default)\033[0m
-  \033[1marm32\033[0m        32-bit ARM \033[90m(Requires arm-none-eabi-gcc)\033[0m
+\x1b[1;36mArchitectures:\x1b[0m
+  \x1b[1mx86_32\x1b[0m       32-bit x86 \x1b[90m(default)\x1b[0m
+  \x1b[1marm32\x1b[0m        32-bit ARM \x1b[90m(Requires arm-none-eabi-gcc)\x1b[0m
+
+\x1b[1;36mFlags:\x1b[0m
+  \x1b[1m--log\x1b[0m        Write all command output text to \x1b[90mbuild.log\x1b[0m
 """